requirements.txt
torchmetrics
datasets
torch
transformers
matplotlib
torchvision
accelerate
numpy
tqdm
Pillow
requests